+++ package.json
@@ -29,6 +29,7 @@
         "vuex": "^3.1.0"
     },
     "dependencies": {
+        "jsbarcode": "^3.11.0",
         "tempusdominus-bootstrap-4": "^5.1.2"
     }
 }
